services/chat-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-genai==1.14.0
python-dotenv==1.1.0
aiosqlite==0.21.0
pydantic==2.11.3
PyJWT==2.12.1
bcrypt==5.0.0
